Distributed w2wHelper functions to helpers/w2w/
