+++ src/components/chat/w2wChat/messageFeed/MessageFeed.tsx
@@ -13,17 +13,18 @@ import ChatSnap from 'components/chat/chatsnap/ChatSnap';
 import LoaderSpinner, { LOADER_TYPE } from 'components/reusables/loaders/LoaderSpinner';
 import { ItemVV2, SpanV2 } from 'components/reusables/SharedStylingV2';
 import { ethers } from 'ethers';
-import { walletToCAIP10 } from 'helpers/w2w';
+import {  decryptFeeds, walletToCAIP10 } from 'helpers/w2w';
 import useToast from 'hooks/useToast';
 import { checkConnectedUser } from 'helpers/w2w/user';
 import { AppContext, Context } from 'sections/chat/ChatMainSection';
 import { MdError } from 'react-icons/md';
 import { intitializeDb } from '../w2wIndexeddb';
-import { decryptFeeds,fetchInbox } from 'helpers/W2WHelper';
+import { fetchInbox } from 'helpers/w2w/ipfs';
 import './MessageFeed.css';
 
 // Internal Configs
 import GLOBALS from 'config/Globals';
+
 
 interface MessageFeedProps {
   filteredUserData: User[];
