+++ src/helpers/w2w/index.ts
@@ -2,6 +2,7 @@ import * as PGP from './pgp'
 import * as DIDHelper from './did'
 import * as Ceramic from './ceramic'
 import * as AES from './aes'
+import { ConnectedUser, Feeds } from 'api'
 
 export const walletToCAIP10 = ({ account, chainId }: { account: string; chainId: number }): string => {
   if (account.includes('eip155:')) {
@@ -65,8 +66,44 @@ export const decryptAndVerifySignature = async ({
   return AES.decrypt({ cipherText, secretKey })
 }
 
+export const decryptFeeds = async ({
+  feeds,
+  connectedUser,
+}: {
+  feeds: Feeds[]
+  connectedUser: ConnectedUser
+}): Promise<Feeds[]> => {
+  for (let feed of feeds) {
+    if (feed.msg.encType !== 'PlainText' && feed.msg.encType !== null) {
+      // To do signature verification it depends on who has sent the message
+      let signatureValidationPubliKey: string
+      if (feed.msg.fromCAIP10 === connectedUser.wallets.split(',')[0]) {
+        signatureValidationPubliKey = connectedUser.publicKey
+      } else {
+        signatureValidationPubliKey = feed.publicKey
+      }
+
+      feed.msg.lastMessage = await decryptAndVerifySignature({
+        cipherText: feed.msg.lastMessage,
+        encryptedSecretKey: feed.msg.encryptedSecret,
+        publicKeyArmored: signatureValidationPubliKey,
+        signatureArmored: feed.msg.signature,
+        privateKeyArmored: connectedUser.privateKey
+      })
+    }
+  }
+  return feeds
+}
+
+export const formatFileSize = (size: number): string => {
+  const i = Math.floor(Math.log(size) / Math.log(1024))
+  return `${(size / Math.pow(1024, i)).toFixed(1)} ${['B', 'KB', 'MB', 'GB', 'TB'][i]}`
+}
+
+
 export default {
   PGP: PGP,
   DID: DIDHelper,
   Ceramic: Ceramic
 }
+
